+++ NEWS.md
@@ -1,4 +1,6 @@
-# enrichplot 1.13.0.992
+# enrichplot 1.13.0.993
++ add `lineSize1`, `lineSize2` and `legendkeySize` parameters for `gseaplot2`(2021-6-13, Sun)
++ add `nodeSize` parameter for `upsetplot`
 + add `color_category` and `color_gene` parameters for `cnetplot`(2021-6-11, Fri) 
 + Enables `showCategory` parameter to support character input in `dotplot.compareClusterResult`(2021-6-10, Thu)